backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pydantic[email]==1.10.12
python-jose
cryptography
asyncpg
passlib[bcrypt]
python-dotenv
psycopg2-binary
bcrypt==4.0.1
python-multipart
redis==4.6.0
elasticsearch[async]==8.8.0


